Refactor into its own allowlist module
